refactor: extract helpers from MainWindow
