+++ projectbaluga/projectbaluga.csproj
@@ -125,6 +125,8 @@
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
+    <Compile Include="Helpers\DeviceManagerHelper.cs" />
+    <Compile Include="Helpers\KeyboardHook.cs" />
     <Page Include="PasswordDialog.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
